fixed tests and thread on start server
