--- remote_robot_web/control_board/__init__.py
@@ -1,5 +1,0 @@
-from control_board.socket_connection import SocketConnection
-
-socket_thread = SocketConnection()  # Creates the socket thread to connect in localhost with ROS
-if not socket_thread.is_alive():
-    socket_thread.start()
